refactor(core): shared useThreadSelectionEvents emitter; export ThreadsEvents
