+++ packages/core/src/store/internal.ts
@@ -18,3 +18,4 @@ export {
   suggestionTriggerDisabled,
 } from "./primitive-predicates";
 export { isDevelopment } from "./env";
+export { useThreadSelectionEvents } from "./clients/thread-selection-events";
