+++ packages/core/src/store/runtime-clients/thread-list-runtime-client.ts
@@ -1,11 +1,8 @@
-import { useEffect, useMemo, useRef } from "react";
+import { useMemo } from "react";
 import { useResource, withKey, resource } from "@assistant-ui/tap";
 import type { ClientOutput } from "@assistant-ui/store";
-import {
-  useAssistantEmit,
-  useClientLookup,
-  useClientResource,
-} from "@assistant-ui/store/client";
+import { useClientLookup, useClientResource } from "@assistant-ui/store/client";
+import { useThreadSelectionEvents } from "../clients/thread-selection-events";
 import type { ThreadListRuntime } from "../../runtime/api/thread-list-runtime";
 import type { AssistantRuntime } from "../../runtime/api/assistant-runtime";
 import { useSubscribable } from "./useSubscribable";
@@ -44,19 +41,7 @@ const useThreadListClient = ({
   __internal_assistantRuntime: AssistantRuntime;
 }): ClientOutput<"threads"> => {
   const runtimeState = useSubscribable(runtime);
-  const emit = useAssistantEmit();
-
-  const { mainThreadId } = runtimeState;
-  const previousMainThreadIdRef = useRef(mainThreadId);
-  useEffect(() => {
-    const previousThreadId = previousMainThreadIdRef.current;
-    if (previousThreadId === mainThreadId) return;
-    previousMainThreadIdRef.current = mainThreadId;
-    emit("threads.selectionChanged", {
-      threadId: mainThreadId,
-      previousThreadId,
-    });
-  }, [mainThreadId, emit]);
+  useThreadSelectionEvents(runtimeState.mainThreadId);
 
   const main = useClientResource(
     ThreadClient({
